finished packages.c and made minor chnages to othr files

+++ include/memory.h
@@ -1,6 +1,6 @@
 #ifndef MEMORY_H_
 #define MEMORY_H_
 
-int get_ram_info(long*, long*, long*);
+int get_ram_info(double *, double*, double*);
 
 #endif // MEMORY_H_
